bot/requirements.txt
slack_bolt==1.22.0
slack_sdk==3.33.5
Flask==3.0.3
python-dotenv==1.0.1
mcp[cli]==1.9.4
